requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
requests~=2.25.1